Refactored the code for ClientOptions Class
